requirements.txt
pokebase==1.3.0
numpy==1.17.3
Pillow==8.3.2
